expand the page to two dimensions!
images loaded in with y-indexes not only render correctly but up and
down kyebindings work. at the moment movement is restricted but only
when trying to move to a spot without a panel. I had to remove linked
paths between panels to get the configuration to not be a nightmare. I'd
like to add them back in at some point.