refactor kinetic proof-of-concept
this strips most of the obsolete code that used MiniDaemon to animate
frame moving, as kinetic does that much more simply. this includes
replacing the keypress event handling with mouse clicks on each image.

minidaemon.js and panel.js are currently not needed but they prove
useful in the future.