put buttons back
the buttons are back to navigate left/right/up/down. the way the
listeners were attached though required using 'cwine' instead of 'this'
to scope variable in the functions. I think there's another way to
ensure they are locked but I forget the syntax.

currIndex was removed since clicking on a panel does nothing to update
it. I'll have to think of some other kinf of bound checking.