use image offset instead of adjusting everywhere
Kinetic lets you set an offset for images so the center of
rotation/animation/etc can remain unchanged. this also means any
calculations made throughout the code that was adjusting for this
doesn't need to be made because now image.x() and image.y() are at the
image center point.