store path lines within panels
this resurrects the Panel class as a wrapper around Kinetic.Image so
that it can extend the class without polluting it. this adds in an extra
object chain everywhere we refer to the actual image though, so I may
change it to a Factory and just add these properties directly to the
Kinetic.Image objects.